+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   - The returned buffers support `Write` trait, and can be inspected/modified what has been written so far
   - The buffer does not allow any access to "dirty" (unset) portion of the buffer
   - The buffer must be finalized with `finish()`, which returns `VCL_STRING`, `VCL_BLOB`, or `&[T]` depending on the builder used
-- If header is not a valid UTF-8 str, `txt::to_str` and `txt::parse_header` will return `None`. Use `txt::to_slice` to handle raw bytes. No more panics here.
+- Add support for the non-UTF8 header values by using a `StrOrBytes` enum. This also fixes crashes when non-UTF8 headers are present.
 - Remove `vsc` feature - all of its functionality is now available without any feature flags
 - Rename `Stat` &rarr; `Metrics`, `Stats` &rarr; `MetricsReader`, `StatsBuilder` &rarr; `MetricsReaderBuilder`, and `Format` &rarr; `MetricsFormat`
 - `MetricsReaderBuilder::patience` now returns `Self`